+++ controller/userController.js
@@ -15,61 +15,27 @@ const bcrypt =  require('bcrypt')
 const jwt =  require('jsonwebtoken')
 require('dotenv').config()
 
-
-// const handleLogin = (req, res) => {
-//     const { username, password } = req.body
-//     const foundUser =  UsersDB.users.find((user) => user.username === username)
-//     if(!foundUser) res.sendStatus(404)
-//     const match = bcrypt.compareSync(password, foundUser.password); // true
-//     if(!match) res.sendStatus(403)
-
-//     const accessToken = jwt.sign(
-//         { "username": foundUser.username },
-//         process.env.ACCESS_TOKEN_SECRET,
-//         {expiresIn: '30s'}
-//     ) // do not pass in sensitive data like passwords as jwt payload
-    
-//     const refreshToken = jwt.sign(
-//         { "username": foundUser.username },
+// const genAccessToken = (req, res) => {
+//     const cookies =  req.cookies
+//     if(!cookies?.jwt) return res.sendStatus(401)
+//     const refreshToken =  cookies.jwt
+//     const foundUser =  UsersDB.users.find((user) => user.refreshToken === refreshToken)
+//     if(!foundUser) return res.sendStatus(403)
+//     // console.log(foundUser)
+//     jwt.verify(
+//         {"username": foundUser.username},
 //         process.env.REFRESH_TOKEN_SECRET,
-//         {expiresIn: '1d'}
-//     ) 
-
-//     //update the foundUser with a refreshToken
-//     UsersDB.setUsers([
-//         {...foundUser, refreshToken}, 
-//         ...UsersDB.users.filter((user) => user.username !== foundUser.username)
-//     ])
-//     /* 
-//         return the token to the front & remeber to keep it in memory / as httpOnly cookie wc is not 
-//         accessible by js
-//         A cookie is always sent with every request but httpOnly cookies are not accesible to js
-//     */
-//     res.cookie('jwt', refreshToken, {httpOnly: true, maxAge: 24*60*60*1000})
-//     res.status(200).json({accessToken})
+//         (err, decoded) => {
+//             if(err || foundUser.username !== decoded.username) return res.sendStatus(403)
+//             const accessToken =  jwt.sign(
+//                 {"username": decoded.username},
+//                 process.env.ACCESS_TOKEN_SECRET,
+//                 { expiresIn: '30s' }
+//             )
+//             res.json({ accessToken })
+//         }
+//     );
 // }
-
-const genAccessToken = (req, res) => {
-    const cookies =  req.cookies
-    if(!cookies?.jwt) return res.sendStatus(401)
-    const refreshToken =  cookies.jwt
-    const foundUser =  UsersDB.users.find((user) => user.refreshToken === refreshToken)
-    if(!foundUser) return res.sendStatus(403)
-    // console.log(foundUser)
-    jwt.verify(
-        {"username": foundUser.username},
-        process.env.REFRESH_TOKEN_SECRET,
-        (err, decoded) => {
-            if(err || foundUser.username !== decoded.username) return res.sendStatus(403)
-            const accessToken =  jwt.sign(
-                {"username": decoded.username},
-                process.env.ACCESS_TOKEN_SECRET,
-                { expiresIn: '30s' }
-            )
-            res.json({ accessToken })
-        }
-    );
-}
 
 const handleLogout = (req, res) => {
     // delete access token on client
